backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain==0.0.335
openai==1.3.0
pinecone-client==2.2.4
supabase==2.0.3
redis==5.0.1
python-multipart==0.0.6
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-socketio==5.10.0
aiohttp==3.9.1
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2 